requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic-settings==2.9.1
python-multipart==0.0.20

# OpenAI and AI dependencies
openai==1.88.0
langchain==0.3.25
langchain-openai==0.3.23
sentence-transformers==4.1.0
transformers==4.52.4
openai-agents-sdk==0.1.2

# Vector database
qdrant-client==1.14.3
pgvector==0.4.1
psycopg2-binary==2.9.10


# Database and ORM
sqlalchemy==2.0.41
alembic==1.16.2
asyncpg==0.30.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.11.7
python-docx==1.2.0
unstructured==0.17.2
PyMuPDF==1.23.8

# NLP and text processing
spacy==3.7.4
pl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_sm-3.7.0/pl_core_news_sm-3.7.0-py3-none-any.whl
nltk==3.8.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1
redis==6.2.0
celery==5.5.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiofiles==25.1.0

# Development and testing
pytest==8.4.0
pytest-asyncio==1.0.0
black==25.1.0
ruff==0.11.13
mypy==1.16.1

# Monitoring and logging
structlog==25.4.0
sentry-sdk==2.30.0

# Progress bars and utilities
tqdm==4.67.1
SQLModel==0.0.24
debugpy==1.8.14

# Added dependency
email-validator>=2.0
alembic>=1.16.2
dateparser
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0

openai==1.97.0
openai-agents==0.2.2
